# Some warnings fixed.
<==== Mercurial data ====>
Branch: vms_3.1_dev

+++ common_libs/nx_utils/src/nx/utils/db/detail/detail_db_structure_updater.cpp
@@ -90,7 +90,7 @@ void DbStructureUpdater::addFullSchemaScript(
 
 unsigned int DbStructureUpdater::maxKnownVersion() const
 {
-    return m_initialVersion + m_updateScripts.size();
+    return m_initialVersion + (unsigned int) m_updateScripts.size();
 }
 
 void DbStructureUpdater::setVersionToUpdateTo(unsigned int version)
@@ -220,7 +220,7 @@ DBResult DbStructureUpdater::updateDbVersion(
 {
     QSqlQuery updateDbVersionQuery(*queryContext->connection());
     updateDbVersionQuery.prepare(R"sql(
-        REPLACE INTO db_version_data(schema_name, db_version) 
+        REPLACE INTO db_version_data(schema_name, db_version)
         VALUES (:schemaName, :dbVersion)
     )sql");
     updateDbVersionQuery.bindValue(":schemaName", QString::fromStdString(m_schemaName));
